spack.util.archive: extract some tarfile utilities
